Struct hidapi::HidApi

pub struct HidApi { /* fields omitted */ }

Object for handling hidapi context and implementing RAII for it. Only one instance can exist at a time.

Methods

impl HidApi

pub fn new() -> HidResult<Self>

Initializes the hidapi.

Will also initialize the currently available device list.

pub fn refresh_devices(&mut self) -> HidResult<()>

Refresh devices list and information about them (to access them use devices() method)

pub fn devices(&self) -> &Vec<HidDeviceInfo>

Returns list of objects containing information about connected devices

pub fn open(&self, vid: u16, pid: u16) -> HidResult<HidDevice>

Open a HID device using a Vendor ID (VID) and Product ID (PID).

When multiple devices with the same vid and pid are available, then the first one found in the internal device list will be used. There are however no grantees, which device this will be.

pub fn open_serial(&self, vid: u16, pid: u16, sn: &str) -> HidResult<HidDevice>

Open a HID device using a Vendor ID (VID), Product ID (PID) and a serial number.

pub fn open_path(&self, device_path: &CStr) -> HidResult<HidDevice>

The path name be determined by inspecting the device list available with HidApi::devices()

Alternatively a platform-specific path name can be used (eg: /dev/hidraw0 on Linux).